ecomm-django-backend/ecomm_django_backend/requirements.txt
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
django-cors-headers